requirements.txt
sentencepiece
transformers
sentence_transformers
qdrant-client
pandas
tqdm
streamlit
torch-scatter